+++ src/main/java/com/intellecteu/onesource/integration/model/onesource/TradeAgreement.java
@@ -15,9 +15,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.intellecteu.onesource.integration.exception.ValidationException;
 import com.intellecteu.onesource.integration.model.enums.ProcessingStatus;
 import com.intellecteu.onesource.integration.services.reconciliation.Reconcilable;
-import java.math.BigDecimal;
-import com.intellecteu.onesource.integration.model.enums.ProcessingStatus;
-import com.intellecteu.onesource.integration.services.reconciliation.Reconcilable;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -39,7 +36,7 @@ public class TradeAgreement implements Reconcilable {
     private Venue venue;
     private Instrument instrument;
     private Rate rate;
-    private BigDecimal quantity;
+    private Integer quantity;
     private CurrencyCd billingCurrency;
     private Double dividendRatePct;
     private LocalDate tradeDate;
