+++ src/main/java/com/intellecteu/onesource/integration/services/IntegrationModelDataTransformer.java
@@ -45,7 +45,6 @@ import com.intellecteu.onesource.integration.model.onesource.Venue;
 import com.intellecteu.onesource.integration.model.onesource.VenueParty;
 import com.intellecteu.onesource.integration.model.onesource.VenueType;
 import io.micrometer.common.util.StringUtils;
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
@@ -122,7 +121,7 @@ public class IntegrationModelDataTransformer implements IntegrationDataTransform
             .venue(buildExecutionVenueFromPosition(position))
             .instrument(buildInstrumentFromPosition(position))
             .rate(buildRebateRateFromPosition(position))
-            .quantity(BigDecimal.valueOf(position.getQuantity()))
+            .quantity(position.getQuantity().intValue())
             .billingCurrency(CurrencyCd.fromValue(position.getCurrency().getCurrencyKy()))
             .dividendRatePct(position.getLoanBorrow().getTaxWithholdingRate())
             .tradeDate(position.getTradeDate().toLocalDate())
