+++ src/test/java/com/intellecteu/onesource/integration/EntityTestFactory.java
@@ -51,7 +51,6 @@ import com.intellecteu.onesource.integration.repository.entity.onesource.Transac
 import com.intellecteu.onesource.integration.repository.entity.onesource.VenueEntity;
 import com.intellecteu.onesource.integration.repository.entity.onesource.VenuePartyEntity;
 import com.intellecteu.onesource.integration.repository.entity.toolkit.CloudSystemEventEntity;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -95,7 +94,7 @@ public class EntityTestFactory {
             .venue(buildVenueEntity())
             .instrument(buildInstrumentEntity())
             .rate(buildRateEntity())
-            .quantity(BigDecimal.valueOf(2L))
+            .quantity(2)
             .billingCurrency(USD)
             .dividendRatePct(2d)
             .tradeDate(LocalDate.now())
